@@ -3,8 +3,8 @@ package com.parisjug.checkout.contracts;
 import com.parisjug.checkout.domain.Delivery;
 import com.parisjug.checkout.domain.Order;
 import io.restassured.module.webtestclient.RestAssuredWebTestClient;
-import org.junit.Before;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -12,10 +12,10 @@ import org.springframework.cloud.contract.stubrunner.spring.AutoConfigureStubRun
 import org.springframework.cloud.contract.stubrunner.spring.StubRunnerProperties;
 import org.springframework.cloud.contract.verifier.messaging.MessageVerifier;
 import org.springframework.cloud.contract.verifier.messaging.boot.AutoConfigureMessageVerifier;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest
 @AutoConfigureStubRunner
 @AutoConfigureMessageVerifier
@@ -31,7 +31,7 @@ public class ContractsBase {
     @Autowired
     private Delivery delivery;
 
-    @Before
+    @BeforeEach
     public void before() {
         RestAssuredWebTestClient.webTestClient(webTestClient);
     }
