@@ -5,9 +5,9 @@ import com.parisjug.delivery.domain.DeliveryQueue;
 import com.parisjug.delivery.domain.Inventory;
 import com.parisjug.delivery.domain.Order;
 import com.parisjug.delivery.provider.Streams;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -17,15 +17,17 @@ import org.springframework.cloud.contract.stubrunner.spring.AutoConfigureStubRun
 import org.springframework.cloud.contract.stubrunner.spring.StubRunnerProperties;
 import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.messaging.support.MessageBuilder;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import reactor.core.publisher.Mono;
 
 import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest
-public class ProcessOrderTests {
+class ProcessOrderTests {
 
     @Autowired
     private Streams streams;
@@ -35,21 +37,19 @@ public class ProcessOrderTests {
 
     @MockBean
     private Inventory inventory;
-
-    @Before
-    public void before() {
+    
+    @BeforeEach
+    void before() {
         deliveryQueue.clear();
+        Book java = new Book("d4d37e73-77a0-4616-8bd2-5ed983d45d14", "Java", BigDecimal.valueOf(100), 100);
+        Mockito.when(inventory.retrieveBook("d4d37e73-77a0-4616-8bd2-5ed983d45d14")).thenReturn(Mono.just(java));
     }
 
     @Test
-    public void should_correctly_process_order() {
-        String bookId = "d4d37e73-77a0-4616-8bd2-5ed983d45d14";
-        Book book = new Book(bookId, "Java", BigDecimal.valueOf(100), 100);
-        Mockito.when(inventory.retrieveBook(bookId)).thenReturn(Mono.justOrEmpty(book));
-        Order order = new Order(bookId, 2, "yannick");
+    void should_correctly_process_order() {
+        Order order = new Order("d4d37e73-77a0-4616-8bd2-5ed983d45d14", 2, "yannick");
         streams.orders().send(MessageBuilder.withPayload(order).build());
         assertThat(deliveryQueue.ordersInProcess()).containsExactly(order);
-
     }
 
 }
