@@ -2,18 +2,17 @@ package com.parisjug.inventory;
 
 import com.parisjug.inventory.domain.Book;
 import com.parisjug.inventory.domain.BookIdGenerator;
-import com.parisjug.inventory.infra.InMemoryBookInventory;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.parisjug.inventory.provider.InMemoryBookInventory;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.reactive.function.BodyInserters;
 
@@ -22,10 +21,10 @@ import java.math.BigDecimal;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest
 @AutoConfigureWebTestClient
-public class InventoryApplicationTests {
+class InventoryApplicationTests {
 
     @Autowired
     private WebTestClient webTestClient;
@@ -39,20 +38,16 @@ public class InventoryApplicationTests {
     private Book java = new Book("d4d37e73-77a0-4616-8bd2-5ed983d45d14", "Java", BigDecimal.valueOf(100), 100);
     private Book kotlin = new Book("8364948b-6221-4cd8-9fd9-db0d17d45ef8", "Kotlin", BigDecimal.valueOf(120), 20);
 
-    @Before
-    public void before() {
+    @BeforeEach
+    void before() {
         Mockito.when(bookIdGenerator.randomId()).thenReturn("dc8493d6-e2e3-47da-a806-d1e8ff7cd4df");
+        bookInventory.removeAllStocks();
         bookInventory.addBook(java);
         bookInventory.addBook(kotlin);
     }
 
-    @After
-    public void after() {
-        bookInventory.removeAllStocks();
-    }
-
     @Test
-    public void should_create_book_into_inventory() {
+    void should_create_book_into_inventory() {
         Book newBook = Book.builder()
                 .name("Kotlin")
                 .price(BigDecimal.valueOf(120))
@@ -66,7 +61,7 @@ public class InventoryApplicationTests {
                 .build();
 
         webTestClient
-                .post().uri("/books")
+                .post().uri("/v1/books")
                 .accept(MediaType.APPLICATION_JSON)
                 .body(BodyInserters.fromObject(newBook))
                 .exchange()
@@ -75,9 +70,9 @@ public class InventoryApplicationTests {
     }
 
     @Test
-    public void should_retrieve_all_books_from_inventory() {
+    void should_retrieve_all_books_from_inventory() {
         webTestClient
-                .get().uri("/books")
+                .get().uri("/v1/books")
                 .accept(MediaType.APPLICATION_JSON)
                 .exchange()
                 .expectStatus().isOk()
@@ -85,9 +80,9 @@ public class InventoryApplicationTests {
     }
 
     @Test
-    public void should_retrieve_book_from_inventory() {
+    void should_retrieve_book_from_inventory() {
         webTestClient
-                .get().uri("/books/d4d37e73-77a0-4616-8bd2-5ed983d45d14")
+                .get().uri("/v1/books/d4d37e73-77a0-4616-8bd2-5ed983d45d14")
                 .accept(MediaType.APPLICATION_JSON)
                 .exchange()
                 .expectStatus().isOk()
@@ -95,9 +90,9 @@ public class InventoryApplicationTests {
     }
 
     @Test
-    public void should_reduce_stock_from_inventory() {
+    void should_reduce_stock_from_inventory() {
         webTestClient
-                .post().uri("/books/d4d37e73-77a0-4616-8bd2-5ed983d45d14/reduce-stock/2")
+                .post().uri("/v1/books/d4d37e73-77a0-4616-8bd2-5ed983d45d14/reduce-stock/2")
                 .accept(MediaType.APPLICATION_JSON)
                 .exchange()
                 .expectStatus().isOk()
