@@ -3,7 +3,7 @@ package contracts
 org.springframework.cloud.contract.spec.Contract.make {
     request {
         method 'POST'
-        url '/checkout'
+        url '/v1/checkout'
         body("""
         {
             "bookId": "d4d37e73-77a0-4616-8bd2-5ed983d45d14",
