@@ -3,7 +3,7 @@ package contracts
 org.springframework.cloud.contract.spec.Contract.make {
     request {
         method 'GET'
-        url '/books/d4d37e73-77a0-4616-8bd2-5ed983d45d14'
+        url '/v1/v1/books/d4d37e73-77a0-4616-8bd2-5ed983d45d14'
         headers {
             header('Content-Type', 'application/json')
         }
