@@ -1,4 +1,4 @@
-package com.parisjug.inventory.infra;
+package com.parisjug.inventory.provider;
 
 import com.parisjug.inventory.domain.Book;
 import com.parisjug.inventory.domain.BookInventory;
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import javax.annotation.PostConstruct;
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -13,6 +15,12 @@ import java.util.Map;
 public class InMemoryBookInventory implements BookInventory {
 
     private final Map<String, Book> inventory = new HashMap<>();
+
+    @PostConstruct
+    void initDB() {
+        Book java = new Book("d4d37e73-77a0-4616-8bd2-5ed983d45d14", "Java", BigDecimal.valueOf(100), 100);
+        addBook(java);
+    }
 
     @Override
     public Flux<Book> allBooks() {
